Fix: resolve all windows compiler warnings for cross-platform compatibility.
Claude AI to the rescue again...

This commit addresses all compiler warnings when building EvilHack on
Windows with Microsoft Visual C++. These fixes ensure clean compilation
across all platforms without changing game behavior.

Fixed warnings by file:
 * apply.c: Explicit boolean conversion from long (oprops != 0L)
 * cmd.c: Cast time_t to long for elapsed time calculation
          Fix old-style function declaration (wiz_kill)
 * eat.c: Cast drain_level() results to unsigned for comparisons
 * end.c: Cast time_t to long for iso8601_duration()
 * files.c: Suppress unused variable warnings in stub function
 * hack.c: Fix old-style function declaration (ma_break)
 * objnam.c: Cast drain_level() to unsigned, pointer diff to int
 * potion.c: Explicit cast to float for resistance calculation

Platform differences addressed:
 - Windows: long is 32-bit, time_t is 64-bit (requires casts)
 - Linux: both long and time_t are 64-bit (no warnings)
 - MSVC is stricter about implicit conversions than GCC

 All changes are explicit casts or syntax modernization that preserve
 the exact same behavior - no game logic was modified.

+++ src/end.c
@@ -920,7 +920,7 @@ time_t when; /* date+time at end of game */
     putstr(0, ATR_HEADING, "Time information:");
     Sprintf(pbuf, " Play time: %s", iso8601_duration(urealtime.realtime));
     putstr(0, 0, pbuf);
-    Sprintf(pbuf, " Wallclock time: %s", iso8601_duration(getnow() - ubirthday));
+    Sprintf(pbuf, " Wallclock time: %s", iso8601_duration((long) (getnow() - ubirthday)));
     putstr(0, 0, pbuf);
     putstr(NHW_DUMPTXT, 0, "");
